requirements.txt
django==1.11
djangorestframework
Pillow
celery==4.1.0
request
